FIx Query Seats

+++ repository/mysql/passenger.go
@@ -11,10 +11,9 @@ const (
 	date_of_birth, gender, "type", street1, street2, postcode, state, city, country, address_type, 
 	issuing_country, country_of_birth, document_type, nationality, created_at, updated_at 
 	FROM passengers WHERE id=?`
-
 )
 
-func (d *dbReadWriter) ReadPassengerByID(ctx context.Context, id int32) (models.Passenger, error) {
+func (d *dbReadWriter) ReadPassengerByID(ctx context.Context, id int64) (models.Passenger, error) {
 	passenger := models.Passenger{}
 
 	err := d.db.QueryRowContext(ctx, selectPassengerByID, id).Scan(
@@ -43,4 +42,3 @@ func (d *dbReadWriter) ReadPassengerByID(ctx context.Context, id int32) (models.
 
 	return passenger, err
 }
-
